% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummyEncode.R
\name{DummyEncode}
\alias{DummyEncode}
\title{Dummy encode (one-hot encode) your categorical variables}
\usage{
DummyEncode(data, Categorical_Variables)
}
\arguments{
\item{data}{A dense data frame containing all variables to be included in a ridge regression.}

\item{Categorical_Variables}{A vector of variables in the data set the user would like to dummy encode.}
}
\value{
A data frame where the categorical columns are one-hot encoded. Each unique category in the original column maps to a new column in the resulting data frame. If a particular observation was of that category, it will contain a 1 in this column. If not, it will contain a 0.
}
\description{
This function allows the user to one-hot encode categorical variables to include in a ridge regression analysis. The user passes in a data frame and a vector of categorical (Character) columns to be encoded. DummyEncode returns a data frame where the categorical columns are one-hot encoded. Each unique category in the original column maps to a new column in the resulting data frame. If a particular observation was of that category, it will contain a 1 in this column. If not, it will contain a 0.
}
