% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPartRidge.R
\name{PlotPartRidge}
\alias{PlotPartRidge}
\title{Plot the partial residuals obtained with PartRidge}
\usage{
PlotPartRidge(partial_residuals_df, dummy = FALSE, dummy_vars = NULL)
}
\arguments{
\item{partial_residuals_df}{The data frame titled "partial_residuals" located in the output list of the ObtainPartRidge function.}

\item{dummy}{Argument tells PlotPartRidge if the input data set contained any variables that were dummy encoded.}

\item{dummy_vars}{If there are dummy encoded variables, dummy_vars takes in a vector of those variables. Additional pre-processing of dummy encoded partial residual values is necessary for gaining insight into these coefficients.}
}
\value{
An object of class "gg" "ggplot" containing you visualization of the partial residuals of your model coefficients.
}
\description{
This function produces box plots of the data frame of partial residuals.
}
